refactor(stores): game-store 슬라이스 패턴 분리 + hintsUsed 필드 추가
542줄 단일 파일이었던 game-store.ts를 관심사별 슬라이스로 분리:
- GameCoreSlice: 보드/퍼즐/히스토리/게임 라이프사이클
- TimerSlice: 경과 시간/일시정지
- UISlice: 셀 선택/메모 모드/힌트 사용 횟수(신규)

구조:
  stores/
  ├── game-store.ts          (진입점: 슬라이스 조합 + persist)
  ├── types.ts               (공유 타입 + SliceCreator 헬퍼)
  ├── slices/
  │   ├── game-core-slice.ts
  │   ├── timer-slice.ts
  │   └── ui-slice.ts
  └── helpers/
      └── serialization.ts   (직렬화/역직렬화)

- 외부 API(useGameStore, serialize*) 동일 유지 → 기존 컴포넌트 수정 없음
- 342개 테스트 전체 통과, Next.js 빌드 정상
